requirements.txt
qrcode[pil]
pillow
python-dotenv
validators